META: update=AUTO fix b/c of NAs
Old meta-analysis PFT membership files sometimes stored NA's as literal
"NA" strings, which were not being read in as NA's. This caused the
difference checking code to (incorrectly) pick up a difference between
in PFT membership and caused the trait meta analysis to always run.

Here, this is fixed by adding "NA" to the list of strings converted to
NA's when reading old PFT membership. Now, if nothing changes, we don't
re-run the meta-analysis!